requirements.txt
asttokens==3.0.1
basedosdados==2.0.2
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
db-dtypes==1.5.0
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-cloud-bigquery==3.39.0
google-cloud-bigquery-connection==1.19.0
google-cloud-bigquery-storage==2.35.0
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
httplib2==0.31.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
loguru==0.7.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandas-gbq==0.32.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tomlkit==0.11.8
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
